msodraw: fix reading non-complex props after complex ones
For non-complex properties, the situation is easy: there is an array of
record headers, and the value is part of the header as well. The old
code assumed that the complex payload (the size of it is specified by
the value in the header) is right after each header, but it turns out
that the stream first contains all the headers, then the payloads.

Note that in case the only complex property is the last one, then the
behavior is unchanged.

A reproducer for this can be a watermark-enabled doc file produced by
Word, and LO's DffPropSet::ReadPropSet() does the same.