doc: don't try to bytes as text outside Clx
When e.g. ChpxFkp refers to byte ranges defined by offsets, we try to be
helpful and print the matching text, to see a given character property
is applied to what characters. However, it's possible that these offets
refer to picture data -- so only show these bytes as text in case they
are referred as text in Clx. fdo34222-4.doc is a reproducer for this
problem.