make for VBA related storage in an ole container more forgiving
The VBA/Basic dump mode now will search for a vba container within the
passed Ole container. Previously the VBAContainer class needed to get
the VBA root 'DirNode' in it's initialisaton. However... it's better to
search for the VBA subfolder I think, that way we can cater for any
VBA container ( I'm thinking here about the blobs that are stored in ooxml
(I'm not sure what the storage structure of of those ole blobs are ) and
additionally I am thinking about the vba container stored for powerpoint
documents ( those ones are wierd, iirc the vba project is stored as an ole
storage but compressed and stored as a blob of bytes somewhere in one of the
main streams of a powerpoint doc ) In that case if we can assume one can
somehow extract that blob and inflate it ( most likely from some custom
code in the powerpoint binary filter ) then we need to be as flexible as
possibly in trying to parse such a VBA project storage.