Added remaining record readers for dir stream, added licences hdrs to files
Some notes:
a) yes the formatted output is not as good as it should be ( feel free to improve )
  * e.g. GUID it's just bytes outputted no formating at all.
  * nested messages like a REFERENCECONTROL can contain a REFERENCEORIGINAL and
a REFERENCENAME as subrecords. ( this is not reflected in the formatted output )
b) there is a separate record reader for every 'Dir' stream record type, lots of
records are similar, it's sure that better design here would allow more code sharing
to be achieved and the amount of readers could be reduced.
c) probably some interpretation of the information in the ProjectXXXX streams could
be done ( infact we don't yet parse anything out of PROJECTlk ( its just a bunch of
GUIDs with some extra blobs for licence info ) ( doesn't appear in excel files that
I examined so I ignored it for now )

add licences