fix DOCDirStream::getString()
For all the other methods that read numbers, the getFoo variant just
reads the memory, while the readFoo version also adjusts the stream
position. Do the same for strings: make getString() not adjust the
stream position, add a readString() that does so and adapt callers.

With that, we can use the new getString() in DopTypography, and then
seek to the expected position, without trying to figure out what the new
position is.

Also, remove xml mode in globals::getUTF8FromUTF16(): that was used by
docdirstream only, and now even that no longer uses it.