Account for 8bit characters inside TextBytesAtom
The Microsoft documentation does not specify an encoding for TextBytesAtom,
but it says that the bytes are the "low byte of a character in the Unicode
character set with the high byte considered equal to zero". As characters
from the latin1 set are the only ones which would not be changed by this
operation, it seems reasonable to assume that the encoding is actually
cp1252/windows-1252, which is what is used by the new version.