Set up a tentative output selection mechanism.
This is based on an integer parameter, in practise the record id
from pptstream.py:recData, and uses a command-line option to set
the list of record types which should be printed. Also added
command-line option to suppress raw dumpBytes output, and store the
command line option as a global object, as this is the only
practical way to get the values down into the low level routines.