added some record IDs and implemented FeatureData for 0x0868 FEAT
FEAT actually is a collection of records of which FEATHDR specifies the
common data, for example enhanced protection. FEAT contains sequences of
ranges the feature is applied to.

Also, FEAT was wrongly named RANGEPROTECTION (probably from the OOo
documentation).

Added names of IDs in the range 0x086? to 0x088?

See
http://msdn.microsoft.com/en-us/library/dd945945.aspx
http://msdn.microsoft.com/en-us/library/dd907085.aspx
http://msdn.microsoft.com/en-us/library/dd911261.aspx