Handling exception on reading invalid values for the fields
There are situations where a field has an invalid value. In such a
case, an exception is occurs, the program halts execution, and the
output will be incomplete.

With this fix, if a key is not found for the dict, name="INVALID"
will be use in the output XML.

This is an example execution for attachment 171572 from tdf#142021
, "EMF with PolyfillMode set to 0 before the bottom star record"

PolygonFillMode can be either 1 or 2, but the field value has the
value of 0.

PolygonFillMode = {
    0x01: "ALTERNATE",
    0x02: "WINDING"
}

$ /opt/libreoffice7.2/program/python ./emf-dump.py star_pfm0.emf

Traceback (most recent call last):
  File "./emf-dump.py", line 29, in <module>
    main(sys.argv)
  File "./emf-dump.py", line 25, in main
    dumper.dump()
  File "./emf-dump.py", line 20, in dump
    strm.dump()
  File "msodumper/emfrecord.py", line 76, in dump
    handler.dump()
  File "msodumper/emfrecord.py", line 400, in dump
    self.printAndSet("PolygonFillMode", self.readuInt32(), dict=PolygonFillMode)
  File "msodumper/binarystream.py", line 32, in printAndSet
    attrs += ' name="%s"' % dict[value]
KeyError: 0

Change-Id: Id965ba0c47eace3029dd5fa12266f1a144eee41e